monolith_dating_and_matches/requirements.txt
celery==5.2.7
Django==4.0.4
channels==3.0.4
channels-redis==3.4.0
Pillow==9.1.0
psycopg2-binary==2.9.3
django-simple-captcha==0.5.17